- Moved ignore.txt into test folder. Each test folder will now have its own ignore.txt - Modified iteratetests.sh to include N number of folders in case more tests are added under a different category - Modified iteratetests.sh to disable running of tests contained in the ignore.txt in that folder - Modified project2_tests/test18.c and project2_tests/test19.c to include the printf statement - Modified output of test18.c (t18.txt) to be empty. "a" and "b" are both defined before use in test18.c. The result of test19.c is still "main:a,b" - Updated README.md to show use of ignore.txt

+++ iteratetests.sh
@@ -1,35 +1,48 @@
+# Get the location of the cc1 executable
 CC=$1
-echo "Running regression tests"
-rm regression_report.txt
-cd "regression_tests"
-for f in *.c; do
-    testname="${f%.*}"
-    testnum=${testname:4}
-    $CC $f
-    DIFF=`diff t$testnum.txt output.txt`
-    if [[ $DIFF != "" ]]; then
-        echo "Test $testnum failed:" >> ../regression_report.txt
-        echo $DIFF >> ../regression_report.txt
-    fi
-done
 
-cd ..
-echo "Running project 2 tests"
-rm project2_report.txt
-cd "project2_tests"
-for f in *.c; do
-    testname="${f%.*}"
-    testnum=${testname:4}
-    $CC $f
-    DIFF=`diff t$testnum.txt output.txt`
-    if [[ $DIFF != "" ]]; then
-        echo "Test $testnum failed:" >> ../project2_report.txt
-        echo $DIFF >> ../project2_report.txt
+# Add folders to iterate through
+declare -a test_folders=("regression_tests" "project2_tests")
+
+# Run tests
+for folder in "${test_folders[@]}"; do
+    echo "Running tests in ${folder}..."
+    cd $folder
+    report_name="${folder}_report.txt"
+    # Read lines from ignore.txt if it exists
+    if [[ -e ignore.txt ]]; then
+        IFS=$'\n' read -d '' -r -a ignore < ignore.txt
+        echo "Found ignore.txt in ${folder}! Not running tests: ${ignore[*]}"
     fi
+    # Remove previous report if it exists, then create it again in case its empty
+    echo -n "" > ../$report_name
+    # Read each .c file
+    for f in *.c; do
+        testname="${f%.*}"
+        testnum=${testname:4}
+        # Check if test number is in local ignore file
+        # If so, continue to next test
+        if [[ -e ignore.txt ]]; then
+            if [[ " ${ignore[*]} " == *" $testnum "* ]]; then
+                #echo "Ignoring test $testnum..."
+                continue
+            fi
+        fi
+        # Otherwise, run test
+        $CC $f > /dev/null 2>&1
+        DIFF=`diff t$testnum.txt output.txt`
+        if [[ $DIFF != "" ]]; then
+            echo "Test $testnum failed!"
+            echo "Test $testnum failed:" >> ../${report_name}
+            echo $DIFF >> ../${report_name}
+        fi
+    done
+    echo "Report output in $report_name"
+    # Remove output.txt if it exists
+    if [ -f output.txt ]; then
+        rm output.txt
+    fi
+    # Remove all of the GCC generated .s files
+    rm *.s
+    cd ..
 done
-cd ..
-rm output.txt
-rm project2_tests/output.txt
-rm regression_tests/output.txt
-rm project2_tests/*.s
-rm regression_tests/*.s
